Move tag mapping to osm module
