+++ src/main/java/org/opentripplanner/openstreetmap/spi/OSMProvider.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.openstreetmap.spi;
 
 import java.time.ZoneId;
-import org.opentripplanner.graph_builder.module.osm.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.tagmapping.OsmTagMapper;
 import org.opentripplanner.openstreetmap.wayproperty.WayPropertySet;
 
 public interface OSMProvider {
