+++ src/test/java/org/opentripplanner/openstreetmap/tagmapper/NorwayMapperTest.java
@@ -1,10 +1,10 @@
-package org.opentripplanner.graph_builder.module.osm.tagmapper;
+package org.opentripplanner.openstreetmap.tagmapper;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.junit.jupiter.api.Test;
-import org.opentripplanner.graph_builder.module.osm.tagmapping.NorwayMapper;
 import org.opentripplanner.openstreetmap.model.OSMWithTags;
+import org.opentripplanner.openstreetmap.tagmapping.NorwayMapper;
 import org.opentripplanner.openstreetmap.wayproperty.WayPropertySet;
 import org.opentripplanner.street.model.StreetTraversalPermission;
